Introduce AssetDownloader
